phpstan fixes

+++ database/migrations/2020_06_06_222222_create_device_outages_table.php
@@ -24,7 +24,7 @@ return new class extends Migration
             $table->increments('availability_id');
             $table->unsignedInteger('device_id')->index();
             $table->bigInteger('duration');
-            $table->float('availability_perc', 6, 6)->default(0.000000);
+            $table->double('availability_perc')->default(0.000000);
             $table->unique(['device_id', 'duration']);
         });
     }
